Update module to be built with rollup.
