--- output.json.d.ts
@@ -1,4 +1,0 @@
-import {RemoteAreaItem} from "./types.js";
-
-declare const data: RemoteAreaItem[];
-export default data;
